+++ package.json
@@ -55,7 +55,7 @@
     "case": "^1.6.3",
     "chai": "^4.3.10",
     "dotenv": "^16.3.1",
-    "eth-deploy-utils": "1.0.0-beta.3",
+    "eth-deploy-utils": "1.0.0-beta.5",
     "eth-sig-util": "^3.0.1",
     "ethereum-waffle": "^4.0.10",
     "ethereumjs-abi": "^0.6.8",
