+++ scripts/deploy-usd.js
@@ -1,0 +1,28 @@
+require("dotenv").config();
+const hre = require("hardhat");
+const ethers = hre.ethers;
+const path = require("path");
+const EthDeployUtils = require("eth-deploy-utils");
+const { normalize, deployContractViaNickSFactory, keccak256, deployContract } = require("../test/helpers");
+let deployUtils;
+
+const { expect } = require("chai");
+
+async function main() {
+  deployUtils = new EthDeployUtils(path.resolve(__dirname, ".."), console.log);
+
+  await deployUtils.deploy("TetherUSD");
+  await deployUtils.deploy("USDCoin");
+
+  console.log(`
+  
+All deployed. Look at export/deployed.json for the deployed addresses.
+`);
+}
+
+main()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error);
+    process.exit(1);
+  });
